Feat: dps 계산로직 수정 및 소환수 데이터 합산

+++ src/main/kotlin/entity/DpsData.kt
@@ -3,6 +3,6 @@ package com.tbread.entity
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class DpsData(val map:HashMap<String,Int> = HashMap()){
+data class DpsData(val map:HashMap<String,Double> = HashMap()){
 }
